Move shared query types
