+++ source/ProcessManager.Orchestrations.Abstractions/Processes/Shared/BRS_026_028/IActorRequestQueryResult.cs
@@ -14,13 +14,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_026;
-
-// TODO:
-// Should be moved to another namespace because this is shared between BRS 026 + 028.
-// We have talked about combining these BRS's into ne top-folder similar to BRS 023 + 027,
-// and then use subfolders to split them per orchestration OR perhaps even use the same orchestration
-// because their logic is very similar.
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.Shared.BRS_026_028;
 
 /// <summary>
 /// Query result from searching for BRS-026 or BRS-028 orchestration instances.
