+++ source/ProcessManager.Orchestrations.Abstractions/Processes/Shared/BRS_026_028/RequestCalculatedEnergyTimeSeriesResult.cs
@@ -16,13 +16,7 @@ using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationInstance;
 using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_026.V1.Model;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_026;
-
-// TODO:
-// Should be moved to another namespace because this is shared between BRS 026 + 028.
-// We have talked about combining these BRS's into ne top-folder similar to BRS 023 + 027,
-// and then use subfolders to split them per orchestration OR perhaps even use the same orchestration
-// because their logic is very similar.
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.Shared.BRS_026_028;
 
 public record RequestCalculatedEnergyTimeSeriesResult :
     OrchestrationInstanceTypedDto<RequestCalculatedEnergyTimeSeriesInputV1>,
